% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_combine_imputations.R
\name{dw_combine_imputations}
\alias{dw_combine_imputations}
\title{Function used to examine the combined results from all imputed datasets.}
\usage{
dw_combine_imputations(imputeddata, formula, reg_type, outcome_levels,
  imputations)
}
\arguments{
\item{imputeddata}{Imputed dataset.}

\item{formula}{Model formula.}

\item{reg_type}{Type of model. 1 = lm. 2 = lrm}

\item{outcome_levels}{Outcome levels. E.g. "2" if binary.}

\item{imputations}{Number of imputations.}
}
\description{
Function used to examine the combined results from all imputed datasets.
}
\examples{
\dontrun {
dw_combine_imputations
(imputeddata,formula,reg_type,outcome_levels,imputations)
}
}
